refactor: reminder creation and deletion

+++ apps/bot/src/crons/reminders.ts
@@ -1,16 +1,22 @@
 import { botCronJob } from "phase.js"
-
 import { ReminderSchema } from "@repo/schemas"
-
 import { PhaseColour } from "~/utils"
-
 import { EmbedBuilder, GuildTextBasedChannel } from "discord.js"
 
 export default botCronJob("*/5 * * * * *", async (client) => {
+  // Fetch reminders that meet the specified conditions
   const reminders = await ReminderSchema.find({
-    $expr: {
-      $lt: [{ $add: [{ $toLong: "$created" }, "$time"] }, Date.now()],
-    },
+    $or: [
+      {
+        $expr: {
+          $lt: [{ $add: [{ $toLong: "$created" }, "$time"] }, Date.now()],
+        },
+      },
+      {
+        unsent: true,
+        created: { $lt: new Date(Date.now()) },
+      },
+    ],
   })
 
   for (const reminder of reminders) {
@@ -19,19 +25,19 @@ export default botCronJob("*/5 * * * * *", async (client) => {
     ) as GuildTextBasedChannel
 
     if (!channel) {
+      // If the channel doesn't exist, delete the reminder and continue to the next one
       await reminder.deleteOne()
       continue
     }
 
     const createdDate = reminder.created
 
+    // Extract date and time components from the created date
     const year = createdDate.getUTCFullYear()
     const month = ("0" + (createdDate.getUTCMonth() + 1)).slice(-2)
     const day = ("0" + createdDate.getUTCDate()).slice(-2)
     const hours = ("0" + createdDate.getUTCHours()).slice(-2)
     const minutes = ("0" + createdDate.getUTCMinutes()).slice(-2)
-
-    const formattedDate = `${year}/${month}/${day} ${hours}:${minutes}`
 
     await channel
       .send({
@@ -42,16 +48,25 @@ export default botCronJob("*/5 * * * * *", async (client) => {
             : undefined,
         embeds: [
           new EmbedBuilder()
-            .setTitle("Reminder")
+            .setTitle(reminder.name ?? "Reminder")
             .setDescription(reminder.message)
             .setColor(PhaseColour.Primary)
             .setFooter({
-              text: `Created ${formattedDate}`,
+              text: `${reminder.loop ? "Started" : "Created"} ${year}/${month}/${day} ${hours}:${minutes}`,
             }),
         ],
       })
       .catch(() => null)
 
-    await reminder.deleteOne()
+    if (reminder.loop) {
+      // If the reminder is set to loop, mark it as sent and update the created date
+      if (reminder.unsent) reminder.unsent = false
+      reminder.created = new Date()
+
+      await reminder.save()
+    } else {
+      // If the reminder is not set to loop, delete it
+      await reminder.deleteOne()
+    }
   }
 })
